check xc's vs mine

Conflicts:
	pAss.03.master.msa.pl

+++ pAss.01.msa.pl
@@ -1,17 +1,17 @@
 #!/usr/bin/env perl
-use strict;
+
+use Modern::Perl '2014';
+use autodie;
 
 die "usage: $0 query_file blast_file output.dir\n" unless $#ARGV == 2;
 
-my $query = shift;
-my $blast = shift;
-my $out = shift;
+my ($query, $blast, $out) = @ARGV;
 
 die unless -f $query;
 die unless -f $blast;
 
 my $temp = rand().time();
-my $CMD = join('', <DATA>);
+my $CMD = join '', <DATA>;
 $CMD =~ s/blastx.txt/$blast/;
 $CMD =~ s/blastx.rma/$temp.rma/;
 $CMD =~ s/query.fna/$query/;
@@ -19,7 +19,7 @@ $CMD =~ s/example/$out/;
 open CMD, ">$temp" or die;
 print CMD $CMD;
 close CMD;
-mkdir $out;
+mkdir $out unless -d $out;
 
 
 my $i;
@@ -32,9 +32,9 @@ while($signal != 0)
     {
         $scr = int(30000 * rand());
     }
-    unlink "$out.lock";
-    unlink "$out.log";
-    unlink "$out.err";
+    unlink "$out.lock" if -e "$out.lock";
+    unlink "$out.log" if -e "$out.log";
+    unlink "$out.err" if -e "$out.err";
     #$signal = system("xvfb-run -n $scr -f $out.lock -e $out.log ~/temp/megan/MEGAN +g -d -E < $temp");
     $signal = system("xvfb-run -n $scr -f $out.lock -e $out.log ~/local/megan/MEGAN -g -d -E < $temp");
     unless($signal == 0)
